Optimisation of the builder classes
