+++ automation_scripts/builders/config.py
@@ -1,0 +1,4 @@
+CORE_DATABASE = "core_db"
+STAGIC_SCHEMA = "staging"
+TRANSFORMED_SCHEMA = "transformed"
+NOTIFICATION_QUEUE = "AZURE_NOTIFICATION"
